core/requirements.txt
# Core dependencies
pandas>=1.3.0
numpy>=1.20.0
flask>=2.0.0
supabase>=2.15.0
python-dotenv>=0.19.0
dataclasses-json>=0.5.6
pyzmq>=25.1.0

# Supabase dependencies
realtime>=1.0.0
supafunc>=0.3.0
storage3>=0.8.0
postgrest>=0.15.0
httpx>=0.24.0
gotrue>=2.0.0

# Testing dependencies
pytest>=7.0.0
pytest-asyncio>=0.16.0
pytest-cov>=2.12.0

# Development tools
black>=21.5b2
isort>=5.9.1
flake8>=3.9.2

# Documentation
sphinx>=4.0.2
sphinx-rtd-theme>=0.5.2